Rel v0.50.5 (#3332)
* update pulse view

* fix #3301 - pf delete msg

* clean

* short header styles

* fix #3294 event time cols sorting

* fix #3309 label selector fault

* multi arch build

* fix #3328 init co count

* rel notes

+++ internal/tchart/gauge_int_test.go
@@ -1,6 +1,3 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of K9s
-
 package tchart
 
 import (
@@ -11,7 +8,7 @@ import (
 
 func TestComputeDeltas(t *testing.T) {
 	uu := map[string]struct {
-		d1, d2 int64
+		d1, d2 int
 		e      delta
 	}{
 		"same": {
